Updated to 2.0.0 and for consistency with Storyboard extension.
ScreenshotTaker now returns the image dimension rather than width. This is a breaking change.
Deleted test rig classes since these are now in Concordion core. Since we need a couple of additional methods, copies of those are in
the test.concordion.temp package. These are to be removed once the additional methods are added to core. An issue has been raised on
core for this.